% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOPA-methods.R
\name{corrected}
\alias{corrected}
\title{corrected}
\usage{
corrected(object, ...)
}
\arguments{
\item{object}{an object to explain}

\item{...}{ignored}
}
\value{
An ANOPA table with the corrected test statistics.
}
\description{
'corrected()' provides an ANOPA table with only the corrected
statistics.
}
