% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOPA-methods.R
\name{explain}
\alias{explain}
\title{explain}
\usage{
explain(object, ...)
}
\arguments{
\item{object}{an object to explain}

\item{...}{ignored}
}
\value{
a human-readable output with details of computations.
}
\description{
'explain()' provides a human-readable, exhaustive, description of
the results. It also provides references to the key results.
}
