% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOPA-methods.R
\name{summarize}
\alias{summarize}
\title{summarize}
\usage{
summarize(object, ...)
}
\arguments{
\item{object}{an object to summarize}

\item{...}{ignored}
}
\value{
an ANOPA table as per articles.
}
\description{
'summarize()' provides the statistics table an ANOPAobject.
It is synonym of 'summary()' (but as actions are verbs, I used a verb).
}
