% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOPA-methods.R
\name{uncorrected}
\alias{uncorrected}
\title{uncorrected}
\usage{
uncorrected(object, ...)
}
\arguments{
\item{object}{an object to explain}

\item{...}{ignored}
}
\value{
An ANOPA table with the un-corrected test statistics.
That should be avoided, more so if your sample is rather small.
}
\description{
'uncorrected()' provides an ANOPA table with only the uncorrected
statistics.
}
